requirements.txt
absl-py==1.3.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
flatbuffers==22.12.6
gast==0.4.0
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-pasta==0.2.0
googleapis-common-protos==1.57.0
grpcio==1.34.1
gsheet==0.0.1
gspread==5.7.2
h5py==3.1.0
httplib2==0.21.0
idna==3.4
imgaug==0.4.0
keras==2.10.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
libclang==14.0.6
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
pandas==1.5.2
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
shapely==2.0.0
sheet2api==0.1.3
six==1.15.0
termcolor==1.1.0
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.13
wrapt==1.12.1
